pyproject.toml
# File: pyproject.toml
# Task: 0.1.2 - Dependency Management (Poetry Integration)

[tool.poetry]
name = "secure-enterprise-rag"
version = "0.1.0"
description = "Secure Enterprise RAG with PII Masking"
authors = ["Principal Security Architect"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
# Core Python requirement aligned with local M1 environment
python = ">=3.13.1,<3.14"

# API & Server
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.17"
pydantic-settings = "^2.6.0"

# Database & ORM
sqlalchemy = "^2.0.36"
alembic = "^1.13.0"
asyncpg = "^0.30.0"

# Auth & Security Context (Cryptography & Hashing)
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

# Vector DB & Cache
qdrant-client = "^1.12.0"
redis = "^5.2.0"

# Orchestration & LLM
langgraph = "^1.1.10"
langchain-core = "^1.3.2"
langchain-openai = "^1.2.1"
langchain-text-splitters = "^1.1.2"

# Security & PII Masking
presidio-analyzer = "^2.2.35"
presidio-anonymizer = "^2.2.35"

# Document Processing & Observability
pymupdf = "^1.24.0"
structlog = "^24.4.0"
greenlet = "^3.5.0"
argon2-cffi = "^25.1.0"
tenacity = "^9.1.4"
python-docx = "^1.2.0"
sentence-transformers = "^5.4.1"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^1.3.0"
pytest-cov = "^6.0.0"
ruff = ">=0.8.0"
pyright = ">=1.1.389"
pre-commit = "^4.0.0"
httpx = "^0.27.0"
fakeredis = "^2.35.1"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore:Support for class-based `config` is deprecated:DeprecationWarning"
]
markers = [
    "unit: fast isolated unit tests (no I/O, no network)",
    "integration: tests that touch external services (DB, Qdrant, Redis)",
]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = [
    "venv", ".venv", "__pycache__", "build", "dist",
    "logs", "automation", ".pytest_cache", "htmlcov", ".coverage"
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "S", "W", "N", "UP", "C4", "SIM", "RUF", "A", "DTZ", "TID", "ASYNC", "PT"] # Added 'S' (flake8-bandit) for SAST
ignore = ["F401", "E501", "B008", "RUF001", "RUF002", "RUF003", "N815", "S107"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S108"]
"**/tests/*" = ["S101", "S105", "S108"]

[tool.pyright]
include = ["app"]
exclude = ["**/node_modules", "**/.*", "**/__pycache__", "build", "dist", "venv", ".venv", "collect_context.py"]
pythonVersion = "3.13"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportMissingImports = true
extraPaths = ["."]
reportMissingTypeStubs = true      # Force checking for missing stubs
reportUnusedImport = "error"       # Cleanliness standard
reportUnusedVariable = "error"     # Security/Cleanliness

[[tool.pyright.executionEnvironments]]
root = "tests"
extraPaths = ["."]
reportPrivateUsage = "none"
reportUnknownParameterType = "none"
reportMissingParameterType = "none"
reportUnknownArgumentType = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
